Added "Introduction to Applied Cryptography"
Introduction to Applied Cryptography is a course offered by the University of Colorado. From a glance at the syllabus, it seems to be a lot more in-depth than the Cryptography I course. If someone could give the syllabus another glance that would be great, but I think that this would make a nice addition as it includes many of the skills necessary to learn cryptography in the course itself.